requirements.txt
pyloudnorm>=0.1.0  # required for LUFS normalization
numpy>=1.21.0
pydub>=0.25.0
scipy>=1.9.0
pdfplumber>=0.10.3
pytesseract>=0.3.10
fpdf2>=2.7
requests>=2.32.0
faster-whisper>=1.0.0
pyannote.audio>=2.1.1
soundfile>=0.13.1
mido>=1.3.0
torch>=2.0.0
torchaudio>=2.0.0
higgs>=0.0.6
